+++ firebase-firestore/src/main/java/com/google/firebase/firestore/core/FieldFilter.java
@@ -17,15 +17,17 @@ package com.google.firebase.firestore.core;
 import static com.google.firebase.firestore.util.Assert.hardAssert;
 
 import com.google.firebase.firestore.model.Document;
-import com.google.firebase.firestore.model.DocumentKey;
 import com.google.firebase.firestore.model.FieldPath;
 import com.google.firebase.firestore.model.value.ArrayValue;
+import com.google.firebase.firestore.model.value.DoubleValue;
 import com.google.firebase.firestore.model.value.FieldValue;
+import com.google.firebase.firestore.model.value.NullValue;
+import com.google.firebase.firestore.model.value.ReferenceValue;
 import com.google.firebase.firestore.util.Assert;
 import java.util.Arrays;
 
 /** Represents a filter to be applied to query. */
-public class RelationFilter extends Filter {
+public class FieldFilter extends Filter {
   private final Operator operator;
 
   private final FieldValue value;
@@ -36,7 +38,7 @@ public class RelationFilter extends Filter {
    * Creates a new filter that compares fields and values. Only intended to be called from
    * Filter.create().
    */
-  RelationFilter(FieldPath field, Operator operator, FieldValue value) {
+  protected FieldFilter(FieldPath field, Operator operator, FieldValue value) {
     this.field = field;
     this.operator = operator;
     this.value = value;
@@ -55,50 +57,67 @@ public class RelationFilter extends Filter {
     return value;
   }
 
-  @Override
-  public boolean matches(Document doc) {
-    if (this.field.isKeyField()) {
-      Object refValue = value.value();
-      hardAssert(
-          refValue instanceof DocumentKey, "Comparing on key, but filter value not a DocumentKey");
-      hardAssert(
-          operator != Operator.ARRAY_CONTAINS
-              && operator != Operator.ARRAY_CONTAINS_ANY
-              && operator != Operator.IN,
-          "'" + operator.toString() + "' queries don't make sense on document keys.");
-      int comparison = DocumentKey.comparator().compare(doc.getKey(), (DocumentKey) refValue);
-      return matchesComparison(comparison);
-    } else {
-      FieldValue value = doc.getField(field);
-      return value != null && matchesValue(doc.getField(field));
-    }
-  }
-
-  private boolean matchesValue(FieldValue other) {
-    if (operator == Operator.ARRAY_CONTAINS) {
-      return other instanceof ArrayValue && ((ArrayValue) other).getInternalValue().contains(value);
+  /**
+   * Gets a Filter instance for the provided path, operator, and value.
+   *
+   * <p>Note that if the relation operator is EQUAL and the value is null or NaN, this will return
+   * the appropriate NullFilter or NaNFilter class instead of a FieldFilter.
+   */
+  public static FieldFilter create(FieldPath path, Operator operator, FieldValue value) {
+    if (path.isKeyField()) {
+      if (operator == Operator.IN) {
+        hardAssert(
+            value instanceof ArrayValue,
+            "Comparing on key with IN, but an array value was not a RefValue");
+        return new KeyFieldInFilter(path, (ArrayValue) value);
+      } else {
+        hardAssert(
+            value instanceof ReferenceValue,
+            "Comparing on key, but filter value not a ReferenceValue");
+        hardAssert(
+            operator != Operator.ARRAY_CONTAINS && operator != Operator.ARRAY_CONTAINS_ANY,
+            operator.toString() + "queries don't make sense on document keys");
+        return new KeyFieldFilter(path, operator, (ReferenceValue) value);
+      }
+    } else if (value.equals(NullValue.nullValue())) {
+      if (operator != Filter.Operator.EQUAL) {
+        throw new IllegalArgumentException(
+            "Invalid Query. You can only perform equality comparisons on null (via "
+                + "whereEqualTo()).");
+      }
+      return new FieldFilter(path, operator, value);
+    } else if (value.equals(DoubleValue.NaN)) {
+      if (operator != Filter.Operator.EQUAL) {
+        throw new IllegalArgumentException(
+            "Invalid Query. You can only perform equality comparisons on NaN (via "
+                + "whereEqualTo()).");
+      }
+      return new FieldFilter(path, operator, value);
+    } else if (operator == Operator.ARRAY_CONTAINS) {
+      return new ArrayContainsFilter(path, value);
     } else if (operator == Operator.IN) {
-      hardAssert(value instanceof ArrayValue, "'in' filter has invalid value: " + value);
-      return ((ArrayValue) value).getInternalValue().contains(other);
+      hardAssert(value instanceof ArrayValue, "IN filter has invalid value: " + value.toString());
+      return new InFilter(path, (ArrayValue) value);
     } else if (operator == Operator.ARRAY_CONTAINS_ANY) {
       hardAssert(
-          value instanceof ArrayValue, "'array_contains_any' filter has invalid value: " + value);
-      if (other instanceof ArrayValue) {
-        for (FieldValue val : ((ArrayValue) other).getInternalValue()) {
-          if (((ArrayValue) value).getInternalValue().contains(val)) {
-            return true;
-          }
-        }
-      }
-      return false;
+          value instanceof ArrayValue,
+          "ARRAY_CONTAINS_ANY filter has invalid value: " + value.toString());
+      return new ArrayContainsAnyFilter(path, (ArrayValue) value);
     } else {
-      // Only compare types with matching backend order (such as double and int).
-      return value.typeOrder() == other.typeOrder()
-          && matchesComparison(other.compareTo(this.value));
+      return new FieldFilter(path, operator, value);
     }
   }
 
-  private boolean matchesComparison(int comp) {
+  @Override
+  public boolean matches(Document doc) {
+    FieldValue other = doc.getField(field);
+    // Only compare types with matching backend order (such as double and int).
+    return other != null
+        && value.typeOrder() == other.typeOrder()
+        && this.matchesComparison(other.compareTo(value));
+  }
+
+  protected boolean matchesComparison(int comp) {
     switch (operator) {
       case LESS_THAN:
         return comp < 0;
@@ -111,7 +130,7 @@ public class RelationFilter extends Filter {
       case GREATER_THAN_OR_EQUAL:
         return comp >= 0;
       default:
-        throw Assert.fail("Unknown operator: %s", operator);
+        throw Assert.fail("Unknown FieldFilter operator: %s", operator);
     }
   }
 
@@ -138,10 +157,10 @@ public class RelationFilter extends Filter {
 
   @Override
   public boolean equals(Object o) {
-    if (o == null || !(o instanceof RelationFilter)) {
+    if (o == null || !(o instanceof FieldFilter)) {
       return false;
     }
-    RelationFilter other = (RelationFilter) o;
+    FieldFilter other = (FieldFilter) o;
     return operator == other.operator && field.equals(other.field) && value.equals(other.value);
   }